requirements.txt
mysql-connector
Pillow
python-dotenv